backend/brainwave_app/requirements.txt
tensorflow
streamlit
opencv-python-headless
plotly
pandas
numpy
fastapi
uvicorn[standard]
pillow
